fix(master,sdk,preload,sdk): fix some bugs found in security code-scanning

+++ preload/sdk/preloadsdk.go
@@ -41,7 +41,7 @@ import (
 type LimitParameters struct {
 	TraverseDirConcurrency int64
 	PreloadFileConcurrency int64
-	ReadBlockConcurrency   int64
+	ReadBlockConcurrency   int32
 	PreloadFileSizeLimit   int64
 	ClearFileConcurrency   int64
 }
